Prevent app from switching to English on every open
The logic in Splash.onCreate() was broken.  After checking if this is firstLaunch, it should goToMainActivity() w/o doing anything else if this is not the first launch.  The following instructions needed to be in an else block otherwise they would be executed b/c the intent/transition was asynchronous in nature

I kept the Crittercism initialization above the if/else b/c I believe it needs to happen each time