rework cueing the video to address potential race conditions
The `onReady()` callback is only ever fired once. If an updated overview video is returned from the API after `onReady()` is fired the overview video won't be updated.
The `getYouTubePlayerWhenReady()` method will more reliably execute the `cueVideo` command.
In order to prevent the video from reinitializing due to other changing data we utilize the provided `YouTubePlayerTracker` to track which video is currently playing. We only cue a video if it's not the current video.