directly observe the flow instead of using an intermediary LiveData
This fixes a bug where the active page was lost when the activity was recreated. This bug was
caused by the adapter being set before the initial pages were set on the adapter. By using the
flow directly we will immediately have access to the current set of pages when creating the adapter.